+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Allow overriding alchemy-field title & description in extra widget settings
 * Allow overriding an alchemy-field's used wrapper-view
 * Add `AlchemyField#applyOptions(options)` method
+* Allow an alchemy-field's field_title property to be set
 
 ## 0.1.11 (2022-07-23)