+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Allow overriding an alchemy-field's used wrapper-view
 * Add `AlchemyField#applyOptions(options)` method
 * Allow an alchemy-field's field_title property to be set
+* Allow turning off specific column filters in alchemy-table
 
 ## 0.1.11 (2022-07-23)